refactoring+fix bug

+++ src/main/java/carTravel/repository/car/CarRepository.java
@@ -3,7 +3,6 @@ package carTravel.repository.car;
 import carTravel.entity.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 @Repository
-public interface CarRepository extends JpaRepository<Car, Long> {
+public interface CarRepository extends JpaRepository<Car, Integer> {
 }
